refactor: Main validations function - #9
This commit adds the following changes:

Creates the validations.js function responsible for managing the main validations of the script.
The following functions are now handled within the validations function: validateDirectoryExists, validateTimestamps, validateMaxDuration, helpGetVideoDuration, formatTime.
Refactors part of the logic and error handling in the validations managed by the validations function, allowing for better information about the error that occurred, its type, and its origin, including the execution tree.